Load more pagination
Working proof of concept for now - would like to improve it at some point so that only new events are loaded into the events array on each call. So far I have played around with customizing the load function response based on the provided limit search param, but reloading the whole page with such param will not load the events that would otherwise be there on initial load.